python/servers/air-quality/pyproject.toml
[project]
name = "air-quality-server"
version = "0.1.0"
description = "한국환경공단 에어코리아 대기오염정보 MCP 서버"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.3.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "data-go-kr-shared",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
air-quality-server = "src.server:main"

[tool.hatch.metadata]
allow-direct-references = true

[project.optional-dependencies]
dev = []

[tool.uv.sources]
data-go-kr-shared = { path = "../../shared" }
